+++ package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "@xipkg/icons": "^0.9.3",
     "@xipkg/input": "^0.1.1",
+    "@xipkg/select": "^0.2.0",
     "lodash": "^4.17.21",
     "pkg.module.announces": "^0.0.0"
   }
