+++ packages/pkg.module.videoconference/package.json
@@ -12,6 +12,7 @@
     "@livekit/components-styles": "^1.0.7",
     "@livekit/components-core": "^0.8.3",
     "@xipkg/aspect-ratio": "^0.1.2",
+    "@xipkg/select": "0.2.0",
     "@xipkg/icons": "0.8.4",
     "@xipkg/button": "1.1.3",
     "livekit-client": "^1.15.4",
